+++ js/core/draft.js
@@ -48,17 +48,18 @@ define(["db", "globals", "ui", "core/finances", "core/player", "core/team", "uti
      * This is called after draft classes are moved up a year, to create the new UNDRAFTED_3 class. It's also called 3 times when a new league starts, to create all 3 draft classes.
      *
      * @memberOf core.draft
+     * @param {IDBTransaction|null} ot An IndexedDB transaction on players, readwrite; if null is passed, then a new transaction will be used.
      * @param {number} tid Team ID number for the generated draft class. Should be g.PLAYER.UNDRAFTED, g.PLAYER.UNDRAFTED_2, or g.PLAYER.UNDRAFTED_3.
      * @param {number?} scoutingRank Between 1 and 30, the rank of scouting spending, probably over the past 3 years via core.finances.getRankLastThree. If null, then it's automatically found.
      * @param {function()} cb Callback function.
      */
-    function genPlayers(tid, scoutingRank, cb) {
+    function genPlayers(ot, tid, scoutingRank, cb) {
         var withScoutingRank;
 
         withScoutingRank = function (scoutingRank) {
             var agingYears, baseAge, baseRating, draftYear, i, p, playerStore, pot, profile, profiles, tx;
 
-            tx = g.dbl.transaction("players", "readwrite");
+            tx = db.getObjectStore(ot, "players", null, true);
             playerStore = tx.objectStore("players");
 
             profiles = ["Point", "Wing", "Big", "Big", ""];
@@ -92,13 +93,19 @@ define(["db", "globals", "ui", "core/finances", "core/player", "core/team", "uti
                 playerStore.put(p);
             }
 
-            tx.oncomplete = function () {
+            if (ot !== null) {
+                // This function doesn't have its own transaction, so we need to call the callback now even though the update might not have been processed yet.
                 cb();
-            };
+            } else {
+                // This function has its own transaction, so wait until it finishes before calling the callback.
+                tx.oncomplete = function () {
+                    cb();
+                };
+            }
         };
 
         if (!scoutingRank) {
-            g.dbl.transaction("teams").objectStore("teams").get(g.userTid).onsuccess = function (event) {
+            db.getObjectStore(ot, "teams", "teams").get(g.userTid).onsuccess = function (event) {
                 var t;
 
                 t = event.target.result;