+++ CHANGES.md
@@ -17,6 +17,7 @@ number or not.
 - Finals MVP award
 - Season totals, per 36 minutes stats, and career stats are viewable from the main player stats page
 - Option to delete old game data to improve performance
+- Future draft classes are visible up to 3 years in the future
 
 ## Technical details